Favourites implemented with UI elements
This includes the UI stuff for the favourites section. Also included next and previous links to pokemon info page. To Do: Research how to route to the same page(or how to update/recall props) without the need for a clone page to route to. For some reason componentDidMount() was only changing the page on double click not single click